pyproject.toml
[project]
name = "RestaurantBot"
version = "0.1.0"
description = "OpenAI SDK, Assignment"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.3.2",
    "openai-agents[voice]>=0.2.8",
    "python-dotenv>=1.1.1",
    "sounddevice>=0.5.2",
    "streamlit>=1.48.1",
]